requirements.txt
numpy~=1.18.1
tensorflow-gpu~=2.3.0
matplotlib~=3.1.3
tqdm~=4.47.0
wandb~=0.10.29
dataclasses~=0.8
gym~=0.17.1
pygame~=1.9.6
ple~=0.0.1
